Better work out timing patterns
One remaining question is what "availability" scope should mean. I
thinked Scoped.scope is right the way it is. We could have availability
be like a property change listener which would see the change. I think
that would be sensible.

Otherwise it is equivalent to the "scope that determines the option's
availability" which is different from "scope from options availability."