Try to simplify things even further after finding bug
Okay so every time I pull this up after a few months I find something to
rework.

This time it didn't work. Some last refactoring I did to make publishing
events from scoped elements more consistent broke things. It is still
broken.

The problem is that when an element is added that is already in scope,
there is no event published for the UI (or anything else) to listen to.
In this commit, an event is published, but it is published before
anything can listen to it.

I'm also experimenting with "just" using EventStreams in place of
Events, which is basically the same thing except wraps publishing in a
Future. That (publishing in a future) actually "fixes" this bug, however
it is inconsistent. A state change has happened, and then in a later
event loop cycle it is published. Everything in between will see the
state change before those event listeners. Things that use Observable
work more intuitively. Observable in general is a very simple and useful
abstraction. Both Events and Observable are backed by EventStream.

Originally I believe the only reason I ever published things in futures
was to get around the fact that I was using synchronous broadcast
streams for everything and couldn't reenter. Since then (since take 3?)
I've used more conventional async streams, and only using sync for very
specific (but common) use cases.

I'm thinking by switching to use EventStream, the abstractions are
simpler. With simpler abstractions, I should be able to design something
that avoids bugs like the above.