Create child streams inside elements
This is because whether there is a child stream or not doesn't matter to
ScopedElements. It just needs the events to be published to the stream
it provides. This can be done with or without a child stream.

A child stream is only needed when the _element's_ API has an event
stream for only that element. In this case, the element can create a
child stream and maintain its constructor contract while also providing
an element-local stream.